[erp] шаблон поиска изображений для бренда

+++ erp-logics/src/main/java/lsfusion/erp/integration/image/DefaultImageArticleActionProperty.java
@@ -18,6 +18,8 @@ import java.io.*;
 import java.net.URL;
 import java.net.URLConnection;
 import java.sql.SQLException;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class DefaultImageArticleActionProperty extends DefaultIntegrationActionProperty {
 
@@ -42,39 +44,58 @@ public class DefaultImageArticleActionProperty extends DefaultIntegrationActionP
 
         try {
 
+            String patternImageArticle = trim((String) findProperty("patternImageArticle").read(context, articleObject));
             String idArticle = trim((String) findProperty("idArticle").read(context, articleObject), "");
+            if (patternImageArticle != null && idArticle.matches(patternImageArticle + ".*")) {
+                Pattern p = Pattern.compile(patternImageArticle);
+                Matcher m = p.matcher(idArticle);
+                if (m.find()) {
+                    idArticle = m.group(0);
+                }
+            } 
+            idArticle = idArticle.replace("/", "+");
+            
             String idBrandArticle = trim((String) findProperty("idBrandArticle").read(context, articleObject), "");
             String siteBrandArticle = trim((String) findProperty("siteBrandArticle").read(context, articleObject));
             String url = "https://ajax.googleapis.com/ajax/services/search/images?v=1.0&q=" + 
                     idBrandArticle + "%20" + idArticle + "&rsz=" + pageSize + "&start=" + start * pageSize + 
                     (siteBrandArticle==null ? "" : "&as_sitesearch=" + siteBrandArticle);
+            url = url.replace(" ", "+");
 
-            final JSONObject response = JsonReader.read(url);
+            JSONObject response;
+            try {
+                response = JsonReader.read(url);
+            } catch (IOException e) {
+                response = null;
+            }
             if (response != null) {
-                JSONArray objectCollection = response.getJSONObject("responseData").getJSONArray("results");
+                //может вернуть response status 503
+                if (response.get("responseStatus").equals(200)) {
+                    JSONArray objectCollection = response.getJSONObject("responseData").getJSONArray("results");
 
-                for (int i = 0; i < objectCollection.length(); i++) {
+                    for (int i = 0; i < objectCollection.length(); i++) {
 
-                    JSONObject jsonObject = objectCollection.getJSONObject(i);
-                    String thumbnailUrl = jsonObject.getString("tbUrl");
-                    String imageUrl = jsonObject.getString("url");
-                    String width = jsonObject.getString("width");
-                    String height = jsonObject.getString("height");
+                        JSONObject jsonObject = objectCollection.getJSONObject(i);
+                        String thumbnailUrl = jsonObject.getString("tbUrl");
+                        String imageUrl = jsonObject.getString("url");
+                        String width = jsonObject.getString("width");
+                        String height = jsonObject.getString("height");
 
-                    File file = readImage(thumbnailUrl);
-                    if (file != null && imageUrl != null) {
-                        DataObject currentObject = new DataObject(start * pageSize + i);
-                        findProperty("thumbnailImage").change(IOUtils.getFileBytes(file), context, currentObject);
-                        findProperty("urlImage").change(imageUrl, context, currentObject);
-                        findProperty("sizeImage").change(width + "x" + height, context, currentObject);
-                        file.delete();
+                        File file = readImage(thumbnailUrl);
+                        if (file != null && imageUrl != null) {
+                            DataObject currentObject = new DataObject(start * pageSize + i);
+                            findProperty("thumbnailImage").change(IOUtils.getFileBytes(file), context, currentObject);
+                            findProperty("urlImage").change(imageUrl, context, currentObject);
+                            findProperty("sizeImage").change(width + "x" + height, context, currentObject);
+                            file.delete();
+                        }
                     }
-                }
-                findProperty("startImage").change(start + 1, context);
-                findProperty("articleImage").change(articleObject, context);
+                    findProperty("startImage").change(start + 1, context);
+                    findProperty("articleImage").change(articleObject, context);
 
-                if (start == 0)
-                    findAction("chooseImageAction").execute(context, articleObject);
+                    if (start == 0)
+                        findAction("chooseImageAction").execute(context, articleObject);
+                }
             }
         } catch (Exception e) {
             throw Throwables.propagate(e);
